🚚 [Move]: #59 src/lib/tauri/ 配下を {Name}/index.ts のフォルダ化形式に揃える
CLAUDE.md の「コンポーネント / カスタムフック / src/lib/ 配下のモジュールは
{Name}/index.{ts,tsx} に配置する」規約に合わせ、`taskCommands` /
`tauriError` / `invokeWrapped` をそれぞれフォルダ化。テストも各フォルダ
直下の `__tests__/` に移動した（tauriError は単一カテゴリのため
`index.test.ts`、taskCommands は操作別 5 ファイル）。barrel と内部の
相対 import を新しい階層に追従させた。
