+++ src/lib/tauri/taskCommands/__tests__/taskCommands.openProject.test.ts
