+++ src/lib/tauri/tauriError/__tests__/tauriError.behavior.test.ts
