+++ src/lib/tauri/tauriError/index.ts
